uboot-sign: Avoid symlink and install errors when SPL DTB is missing
Fix potential errors during do_deploy and deploy_spl_dtb when SPL_DTB_BINARY or
SPL_BINARY is not present.

Wrapped install command in deploy_spl_dtb() with a file existence check.
Improved condition in do_deploy to check for actual existence of
${SPL_DIR}/${SPL_DTB_BINARY} instead of only relying on variable non-emptiness.
Prevents "install: missing destination file operand" and invalid symlink
creation.

(From OE-Core rev: e94c080b163ed92346291e1a227c8c2b329c1e9e)

+++ meta/classes-recipe/uboot-sign.bbclass
@@ -275,7 +275,9 @@ deploy_spl_dtb() {
 	fi
 
 	# For backwards compatibility...
-	install -Dm644 ${SPL_BINARY} ${DEPLOYDIR}/${SPL_IMAGE}
+	if [ -e "${SPL_BINARY}" ]; then
+		install -Dm644 ${SPL_BINARY} ${DEPLOYDIR}/${SPL_IMAGE}
+	fi
 }
 
 do_uboot_generate_rsa_keys() {
@@ -600,7 +602,7 @@ do_deploy:prepend() {
 		ln -sf ${UBOOT_FITIMAGE_IMAGE} ${DEPLOYDIR}/${UBOOT_FITIMAGE_SYMLINK}
 	fi
 
-	if [ "${SPL_SIGN_ENABLE}" = "1" -a -n "${SPL_DTB_BINARY}" ] ; then
+	if [ "${SPL_SIGN_ENABLE}" = "1" -a -e "${SPL_DIR}/${SPL_DTB_BINARY}" ] ; then
 		ln -sf ${SPL_DTB_IMAGE} ${DEPLOYDIR}/${SPL_DTB_SYMLINK}
 		ln -sf ${SPL_DTB_IMAGE} ${DEPLOYDIR}/${SPL_DTB_BINARY}
 		ln -sf ${SPL_NODTB_IMAGE} ${DEPLOYDIR}/${SPL_NODTB_SYMLINK}
